렌더 보간 값 정상화
0 or 1 -> 실수 값

+++ NGP_TermProject/NGPTermProject/NetworkSyncManager.h
@@ -27,7 +27,7 @@ public:
 private:
 	static float offsetAvg;
 	static float rttAvg;
-	int delayTick;
+	float delayTick;
 	uint64_t baseTick;
 	uint64_t baseServerTimestamp;
